CHROMIUM: drm/tegra: Fix host1x syncpt array OOB access
We use syncpt id to get the syncpt infos from host1x syncpt
array. And the id can be specified by userspace, so in kernel
we need to make sure the ID from userspace doesn't beyond
the boundary of syncpt array.

Unfortunately our host1x driver doesn't do the check right so
this patch fixes it.

BUG=chrome-os-partner:59595
TEST=System boots OK

Change-Id: I07d7d00a838e68b934b3b303a90e9e25e3e72045
Signed-off-by: Mark Zhang <markz@nvidia.com>