pinctrl: qcom: Clear status bit on irq_unmask
This commits reverts the commit 3976c7a1eafc ("pinctrl:
qcom: Don't clear status bit on irq_unmask").
Above change causes an unexpected wakeup on the gpios when the
interrupt is enabled. The gpio interrupt status bit is getting set
after the irq is disabled and causing an immediate interrupt after
enablling the irq. So, revert this change to avoid unexpected/spurious
wakeups.

The above change came in as part of the merge commit 3a75d7a94709
("Merge 4.4.59 into android-4.4")

Change-Id: Ie58116c8ac5c81bc007d2c4398f5edfcbf9988d6
Signed-off-by: Srinivas Ramana <sramana@codeaurora.org>