loop: fix concurrent lo_open/lo_release
commit ae6650163c66a7eff1acd6eb8b0f752dcfa8eba5 upstream.

范龙飞 reports that KASAN can report a use-after-free in __lock_acquire.
The reason is due to insufficient serialization in lo_release(), which
will continue to use the loop device even after it has decremented the
lo_refcnt to zero.

In the meantime, another process can come in, open the loop device
again as it is being shut down. Confusion ensues.

Bug: 72867809

CVE-Fixed: CVE-2018-5344

Change-Id: I4ee94fca4ad13e227fbeffd4e2cba458c4d6e5c4
Reported-by: 范龙飞 <long7573@126.com>
Cc: Ben Hutchings <ben.hutchings@codethink.co.uk>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>